airbnb-clone-backend/requirements.txt
Flask==2.3.3
PyMongo==4.5.0 
dnspython==2.4.2 
flask-pymongo==2.3.0
python-dotenv==1.0.0
flask-json-schema==0.0.5
flask-expects-json==1.7.0
flask-cors==1.10.3
cryptography==41.0.3